seeking/requirements.txt
oauth2>=1.9.0,<2.0.0
absl-py>=0.4.1,<0.5.0
tweepy>=3.6.0,<3.7.0
PyYAML>=3.13.0,<3.14.0
google-cloud-datastore>=1.7.0,<1.8.0
google-cloud-logging>=1.6.0,<1.7.0
google-cloud-error-reporting>=0.30.0,<0.31.0
google-cloud-pubsub>=0.37.2,<0.38.0
psq>=0.7.0,<0.8.0
Flask>=1.0.2,<1.1.0
more-itertools>=4.3.0,<4.4.0
gunicorn>=19.9.0,<20.0.0
honcho>=1.0.1,<1.1.0
dataclasses>=0.6.0,<0.7.0
